v0.1.2

+++ src/useCalendar.ts
@@ -136,16 +136,26 @@ export const useCalendar = () => {
       return
 
     } else if (event.status === 'cancelled') {
-      // delete event
-
-      deleteEvent(sync, event, (error, _) => {
-        if (error) {
-          console.log(chalk.red('deletion failed') + ' @ ' + chalk.gray(source) + chalk.red(' -/-> ')  + chalk.gray(sync.target))
-          console.log(error)
-        } else {
-          console.log(chalk.red('--> deleted event') + ' @ ' + chalk.gray(source + ' -> ' + sync.target))
-        }
+      // check if event to delete was created by CalSync
+      const result = await calendar.events.get({
+        calendarId: sync.target,
+        eventId: event.id,
       })
+
+      if (result.data.creator.email === process.env.GOOGLE_API_CLIENT_MAIL) {
+        // delete event
+        deleteEvent(sync, event, (error, _) => {
+          if (error) {
+            console.log(chalk.red('deletion failed') + ' @ ' + chalk.gray(source) + chalk.red(' -/-> ')  + chalk.gray(sync.target))
+            console.log(error)
+          } else {
+            console.log(chalk.red('--> deleted event') + ' @ ' + chalk.gray(source + ' -> ' + sync.target))
+          }
+        })
+      } else {
+        console.log(chalk.blue('skipped deletion') + ' @ ' + chalk.gray(source) + chalk.red(' -/-> ')  + chalk.gray(sync.target))
+        console.log(chalk.bgBlue('Info: ' + 'Event was not created by CalSync'))
+      }
     }
   }
 
@@ -185,7 +195,7 @@ export const useCalendar = () => {
 
     // send events to queue
     for (const event of events) {
-      await queue.add(event.id, { source, sync, event })
+      await queue.add(event.id, { source, sync, event }, { removeOnComplete: true })
       console.log(chalk.gray(`<-- event queued from ${source}`))
     }
   }
@@ -222,7 +232,12 @@ export const useCalendar = () => {
 
     cache[calendarId].nextSyncToken = result.data.nextSyncToken
     saveCache(cache)
-    return result.data.items
+
+    // ignore events that were created by CalSync
+    return result.data.items.filter(event => {
+      if (!event.creator) return true
+      return event.creator.email !== process.env.GOOGLE_API_CLIENT_MAIL
+    })
   }
 
   const isOutdated = (source: CalendarCacheEntry) => {
